Jenkins/jenkins_scenarios/scenario_03_html_reports/requirements.txt
pytest
pytest-html
pytest-json-report
requests
psycopg2-binary
redis
testcontainers
